@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.2.41'
+    ext.kotlin_version = '1.3.20'
 
     repositories {
         jcenter()
@@ -11,24 +11,20 @@ buildscript {
         google()
     }
     dependencies {
-
-        classpath 'com.android.tools.build:gradle:3.1.3'
+        classpath 'com.android.tools.build:gradle:3.3.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
     }
-}
-
-project.ext {
-    supportVersion = "27.1.1"
-    sdkVersion = 27
 }
 
 allprojects {
     repositories {
         jcenter()
         google()
+        maven { url "https://dl.bintray.com/dasar/maven" }
+        maven { url "https://s3.amazonaws.com/repo.commonsware.com" }
+        maven { url "https://jitpack.io" }
+        mavenCentral()
     }
 }
 
