@@ -7,7 +7,7 @@ import android.app.Service;
 import android.app.job.JobParameters;
 import android.app.job.JobService;
 import android.os.Build;
-import android.support.v4.app.NotificationCompat;
+import androidx.core.app.NotificationCompat;
 import android.util.Log;
 
 import org.horaapps.leafpic.data.HandlingAlbums;
