@@ -2,7 +2,7 @@ package org.horaapps.leafpic.about;
 
 import android.content.Context;
 import android.graphics.Typeface;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.ViewGroup;
 import android.widget.LinearLayout;
 
