@@ -1,6 +1,6 @@
 package org.horaapps.leafpic.animations;
 
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.ViewPager;
 import android.view.View;
 
 /**
