@@ -2,7 +2,7 @@ package org.horaapps.leafpic.fragments;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.View;
 
 import org.horaapps.leafpic.data.Media;
