@@ -1,6 +1,6 @@
 package org.horaapps.leafpic.progress;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.util.ArrayList;
 
