@@ -1,6 +1,6 @@
 package org.horaapps.leafpic.timeline;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
