@@ -4,8 +4,8 @@ import android.content.Context;
 import android.graphics.PorterDuff;
 import android.os.CancellationSignal;
 import android.os.IBinder;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.widget.CardView;
+import androidx.appcompat.app.AlertDialog;
+import androidx.cardview.widget.CardView;
 import android.view.View;
 import android.view.inputmethod.InputMethodManager;
 import android.widget.EditText;
