@@ -2,7 +2,7 @@ package org.horaapps.leafpic.util;
 
 import android.content.Context;
 import android.os.Build;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.Html;
 import android.text.Spanned;
 import android.widget.Toast;
