@@ -1,8 +1,8 @@
 package org.horaapps.leafpic.views.themeable;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.CardView;
+import androidx.annotation.Nullable;
+import androidx.cardview.widget.CardView;
 import android.util.AttributeSet;
 
 import org.horaapps.liz.ThemeHelper;
