@@ -1,7 +1,7 @@
 package org.horaapps.leafpic.views.themeable;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.AttributeSet;
 
 import org.horaapps.liz.ThemeHelper;
