@@ -1,7 +1,7 @@
 package org.horaapps.leafpic.views.themeable;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.AttributeSet;
 
 import org.horaapps.liz.ThemeHelper;
@@ -10,7 +10,7 @@ import org.horaapps.liz.Themed;
 /**
  * Created by darken (darken@darken.eu) on 04.03.2017.
  */
-public class ThemedSettingsTitle extends android.support.v7.widget.AppCompatTextView implements Themed {
+public class ThemedSettingsTitle extends androidx.appcompat.widget.AppCompatTextView implements Themed {
 
     public ThemedSettingsTitle(Context context) {
         this(context, null);
